Commit with Joel's code
I went down a rabbit hole and didn't know where to start. So I watched
the same video as Joel and Eric. I went over Joel's code line by line
for about 7 hours to really understand why it worked and why it wouldn't
work for this assignment. When I imported the code into my atom, it did
not work as expected. Pointless to figure out so I'm starting from
scratch.